Implement main infinite logic to find matches.
On calling the start_matchmaker function, it sets
an infinite setInterval function which will process
the matches.

On processing, the match_ids will be delivered to the
respective client through common room.

NOTE: Might have to optimize if room creation is expensive.

The function proceeds to find match only if previous
function has finished the match finding process.
This is achieved by using a lock-type variable
`last_completed`.

After the match_ids are exchanged, they are cleaned up
from the matchList and the socket_id_map.